+++ CHANGES.rst
@@ -10,6 +10,14 @@ Changelog
   Fixes `issue #19 <https://github.com/collective/collective.catalogcleanup/issues/19>`_.
   [maurits]
 
+- Disable CSRF protection.
+  Fixes `issue #17 <https://github.com/collective/collective.catalogcleanup/issues/17>`_.
+  [maurits]
+
+- Abort any transaction changes in dry run mode.
+  There should not be any changes here anyway, but this makes sure.
+  [maurits]
+
 
 1.8.0 (2018-04-30)
 ------------------