+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-1.8.1 (unreleased)
+1.9.0 (unreleased)
 ------------------
 
 - Catch TypeError when getting object for brain.